requirements.txt
streamlit==1.32.0
pandas==2.2.1
requests==2.31.0
matplotlib==3.8.3
folium==0.16.0
Pillow==10.2.0
altair==5.2.0
numpy==1.26.4
geopy==2.4.1
pydeck==0.8.1b0